Lot of stuff had been improved and added.
I just can't keep up. I'm only using git bash and it suck af. Did not use the git control in Netbeans 8.0.1 (Yeah, because Java is the language we we're taught and Netbeans is also so fking outdated af so yeah. Though, it's okay to migrate to the newer version of Netbeans but a lot of stuff is needed to be adjust since my teacher did not say anything about it and I'm too fking late to know it. So, yeah. Suck af. :) )